Install gitlab-clippy on opensuse:tumbleweed
See https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/407

Install gitlab-clippy into /opt/cargo and adjust the PATH environment
so that it can be run from there easily. Do the same for cargo-c on
debian:testing.

Also rename the Dockerfiles for opensuse and fedora as it was done
for debian:testing recently.